sqlexp: Changed CreateSqlTable
git-svn-id: svn://ultimatepp.org/upp/trunk@9019 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/Core/Vcont.hpp
@@ -288,7 +288,7 @@ void Vector<T>::Insert(int q, const Vector& x, int offset, int count) {
 template <class T>
 void Vector<T>::Insert(int i, std::initializer_list<T> init)
 {
-	RawInsert(i, init.size());
+	RawInsert(i, (int)init.size());
 	T *t = vector + i;
 	for(auto q : init)
 		DeepCopyConstruct(t++, q);
